solicitudes recibidas 21-09-2019

+++ src/views/Shop/listSolicitudesRealizadas.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Table, Button } from "reactstrap";
 import IconButton from "@material-ui/core/IconButton";
-import { Delete, Edit, Visibility, SwapHoriz } from "@material-ui/icons";
+import { Delete, Edit, Visibility, Cancel } from "@material-ui/icons";
 import ModalSolicitudes from './ModalSolicitudes.js';
 import { number_format, GetDisabledPermits, getArray } from "../../core/utils";
 import Pagination from '../../components/Pagination';
@@ -25,6 +25,7 @@ class ListSolicitudesRealizadas extends React.Component {
       status: '',
       page: 0,
       rowsPerPage: 10,
+      typeRequest: 'Realizada',
     };
   }
 
@@ -69,20 +70,37 @@ class ListSolicitudesRealizadas extends React.Component {
     }
   }
 
-  deleteRegister = (id, status) => {
-    const message = {
-      title: "Eliminar Registro",
-      info: "¿Esta seguro que desea eliminar este registro?"
-    };
-    this.props.confirm(message, res => {
-      if (res) {
-        if (status === "Pendiente" || status === "Rechazada") {
-          this.props.DeleteRequestMadeAction(id);
-        } else {
-          this.props.alert("warning", "¡La solicitud no puede ser eliminada, su estatus es: "+status+"!");
+  deleteRegister = (option, id, status) => {
+    if(option === 1){
+      const message = {
+        title: "Eliminar Registro",
+        info: "¿Esta seguro que desea eliminar este registro?"
+      };
+      this.props.confirm(message, res => {
+        if (res) {
+          if (status === "Pendiente" || status === "Rechazada") {
+            this.props.DeleteRequestMadeAction(id);
+          } else {
+            this.props.alert("warning", "¡La solicitud no puede ser eliminada, su estatus es: "+status+"!");
+          }
         }
-      }
-    });
+      });
+    }else{
+      const message = {
+        title: "Cancelar Solicitud",
+        info: "¿Esta seguro que desea cancelar esta solicitud?"
+      };
+      this.props.confirm(message, res => {
+        if (res) {
+          if (status === "Pendiente" || status === "Rechazada") {
+            this.props.cancelRequestAction(id);
+          } else {
+            this.props.alert("warning", "¡La solicitud no puede ser cancelada, su estatus es: "+status+"!");
+          }
+        }
+      });
+    }
+    
   }
 
   valorCloseModal = (valor) => {
@@ -176,7 +194,7 @@ class ListSolicitudesRealizadas extends React.Component {
                         <td style={{ 'minWidth': "205px" }}>
                           <div className="float-left" >
                             <IconButton aria-label="Delete"
-                              title="Ver Transferencia"
+                              title="Ver Solicitud"
                               className="iconButtons"
                               onClick={() => { this.openModal(2, data.number, data._id, data.status); }}
                               disabled={detailsDisabled}>
@@ -184,17 +202,25 @@ class ListSolicitudesRealizadas extends React.Component {
                             </IconButton>
 
                             <IconButton aria-label="Delete"
-                              title="Editar Transferencia"
+                              title="Editar Solicitud"
                               className="iconButtons"
                               onClick={() => { this.openModal(3, data.number, data._id, data.status); }}
                               disabled={updateDisabled}>
                               <Edit className="iconTable" />
+                            </IconButton>                           
+
+                            <IconButton aria-label="Delete"
+                              title="Cancelar Solicitud"
+                              className="iconButtons"
+                              onClick={() => { this.deleteRegister(0, data._id, data.status); }}
+                              disabled={deleteDisabled}>
+                              <Cancel className="iconTable" />
                             </IconButton>
 
                             <IconButton aria-label="Delete"
-                              title="Eliminar Transferencia"
+                              title="Eliminar Solicitud"
                               className="iconButtons"
-                              onClick={() => { this.deleteRegister(data._id, data.status); }}
+                              onClick={() => { this.deleteRegister(1, data._id, data.status); }}
                               disabled={deleteDisabled}>
                               <Delete className="iconTable" />
                             </IconButton>
